Move rule fixtures out of examples folder
